When user books experiment, materials are withdrawn from stock. If insufficient stock, experiment is not booked, user returns to planning page, notification shows up.
Also implemented when user changes experiment, stock levels are not widthdrawn twice. Stock is first reset to original level and then new level is withdrawn;
When changing level but insufficient stock, the changes are not done. Stock levels stays the same as level before change.

Deletion of experiment returns amounts to stock.